chore: refactor sql.rs (#7846)
