+++ src/service/search/sql/rewriter/mod.rs
@@ -1,0 +1,23 @@
+// Copyright 2025 OpenObserve Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+pub mod add_new_filter;
+pub mod add_o2_id;
+pub mod add_ordering_term;
+pub mod add_timestamp;
+pub mod approx_percentile;
+pub mod index;
+pub mod remove_dashboard_placeholder;
+pub mod track_total_hits;
