+++ src/service/search/sql/visitor/mod.rs
@@ -1,0 +1,27 @@
+// Copyright 2025 OpenObserve Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+#[cfg(feature = "enterprise")]
+pub mod cipher_key;
+pub mod column;
+#[cfg(feature = "enterprise")]
+pub mod group_by;
+pub mod histogram_interval;
+pub mod index_optimize;
+pub mod match_all;
+pub mod partition_column;
+pub mod pickup_where;
+pub mod prefix_column;
+pub mod utils;
